Move function 'char' to separate file
